feat(tasks): Add use case and tests for adding task to "my_day"
This commit introduces a new use case and corresponding tests for adding a task to the "my_day" feature. The use case allows updating the "my_day" field of a task to true.

The tests cover scenarios where the update of this field to true is successful and when it should not be allowed if the field value is already true.