feat(task): allow updating task name and description
This commit implements the use case for updating a task of a user. The use case allows the user to update the name and description fields of a task. The implementation includes the necessary logic to validate and update the task's name and description in the repository. Also was create a test suite for the use case.